responsive design
this is the homepage for the portfolio site. using a grid structure the main content of the site, the grid/content does shrink when the window is squeezed, but I don't know how to make the grid stack into 4 rows rather than keep the 2x2 structure. And I don't know how to make the 'fan' image responsive